minior tidyup and removing artifact .git repo
